Completely redid the way time works
-No longer use millis()
-No longer have 2 different variables for time but instead use only RTC
-Added GPS OKAY message for testing purposes
Note: Have to test if this works if not then we need to look at other
possibilities for why this program doesnt work

+++ Container Softare/ContinerSoftwareV2/SD.ino
@@ -9,7 +9,7 @@ void setupSD(){
 }
 
 //Writes any string data to SD card
-void writeToSD(double data, char FileName){
+void writeToSD(float data, char FileName){
   File file;
   file = SD.open(String(FileName), FILE_WRITE);
   file.println(data);
